Feat: getSettingsOptions
