+++ packages/blocks/src/utils/from-theme-json/getLayoutOptions.ts
@@ -1,8 +1,7 @@
+import { isEmpty } from 'lodash';
 import type { ThemePropertiesLayout } from '../../types/layout';
 
-export interface LayoutProperties {
-  layout?: Record<string, unknown>;
-}
+export type LayoutProperties = Record<string, unknown>;
 
 /**
  * Takes an object with an optional `layout` property and transforms it into the ThemePropertiesLayout object type.
@@ -10,17 +9,12 @@ export interface LayoutProperties {
  * @returns {ThemePropertiesLayout} An object with type ThemePropertiesLayout
  */
 export function getLayoutOptions(
-  props: LayoutProperties,
+  layout?: LayoutProperties,
 ): ThemePropertiesLayout {
-  if (props?.layout) {
+  if (!isEmpty(layout)) {
     /* eslint-disable @typescript-eslint/no-unsafe-return */
     return {
-      ...((props.layout?.contentSize && {
-        contentSize: props.layout.contentSize,
-      }) as Record<string, unknown>),
-      ...((props.layout?.wideSize && {
-        wideSize: props.layout.wideSize,
-      }) as Record<string, unknown>),
+      ...layout,
     };
   }
   return {};
